Improved artifact collection message (#1028)

+++ rest/src/main/java/software/wings/common/Constants.java
@@ -367,7 +367,7 @@ public interface Constants {
   /**
    * The constant DEFAULT_STATE_TIMEOUT_MILLIS.
    */
-  long DEFAULT_ARTIFACT_COLLECTION_STATE_TIMEOUT_MILLIS = 5 * 60 * 1000; // 5 minutes
+  Integer DEFAULT_ARTIFACT_COLLECTION_STATE_TIMEOUT_MILLIS = 5 * 60 * 1000; // 5 minutes
 
   /**
    * O
